Add RSI bindings to 'new board' input
This time I added the code in-line on the input element to see if it
would work (it did)

This isn't as robust as the version in BoardShow.vue, I want to leave
this here to remind me to research why I can't access the 'window'
object from inside vue events/props.

Also Its worth mentioning the bindings don't work exactly like
Vim/Emacs/term. Tokens delimited by spaces are considered words,
regardless of any symbols or digits (typically '_' is considered a word
character but '-' is not.)

There's a bug I want to fix with C-W not deleting the last word if
there's a space just behind the cursor.